+++ packages/transaction-controller/src/index.ts
@@ -36,7 +36,7 @@ export type {
   TransactionControllerGetTransactionsAction,
   TransactionControllerUpdateCustodialTransactionAction,
   TransactionControllerUpdateTransactionAction,
-  TransactionControllerUpdateTransactionMetadataAction,
+  TransactionControllerUpdateTransactionMetadataWithoutResimulationAction,
   TransactionControllerHandleMethodDataAction,
   TransactionControllerIsAtomicBatchSupportedAction,
   TransactionControllerStopTransactionAction,